Verdict module
