+++ src/message.rs
@@ -34,36 +34,3 @@ impl<'a> Message<'a> {
         }
     }
 }
-
-pub mod verdict {
-    use libc::*;
-    use ffi::*;
-
-    pub enum Verdict {
-        Drop,
-        Accept,
-        Stolen,
-        Queue,
-        Repeat,
-        Stop
-    }
-
-    impl Verdict {
-
-        pub fn set_verdict(qh: *mut nfq_q_handle, packet_id: u32, verdict: Verdict, data_len: u32, buffer: *const c_uchar) -> Result<c_int, ()> {
-            let c_verdict = match verdict {
-                Verdict::Drop => NF_DROP,
-                Verdict::Accept => NF_ACCEPT,
-                Verdict::Stolen => NF_STOLEN,
-                Verdict::Queue => NF_QUEUE,
-                Verdict::Repeat => NF_REPEAT,
-                Verdict::Stop => NF_STOP
-            };
-
-            match unsafe { nfq_set_verdict(qh, packet_id as uint32_t, c_verdict as uint32_t, data_len as uint32_t, buffer) } {
-                -1 => Err(()),
-                r @ _ => Ok(r)
-            }
-        }
-    }
-}
